HUCO-112 | [feat] merge customcategory.cs into category.cs
