--- HandsUp/Assets/CardViewScene/Scripts/Category/CustomCategory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7ae5c673140947d48a12288b93800810
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
